refactor: 쿼리 패키지 변경
