+++ src/main/java/kitchenpos/menus/tobe/ui/MenuRestController.java
@@ -4,9 +4,9 @@ import java.net.URI;
 import java.util.List;
 import java.util.UUID;
 import kitchenpos.menus.tobe.application.MenuCommandHandler;
-import kitchenpos.menus.tobe.application.query.MenuConverter;
-import kitchenpos.menus.tobe.application.query.MenuQueryHandler;
-import kitchenpos.menus.tobe.application.query.result.MenuQueryResult;
+import kitchenpos.menus.tobe.query.MenuConverter;
+import kitchenpos.menus.tobe.query.MenuQueryHandler;
+import kitchenpos.menus.tobe.query.result.MenuQueryResult;
 import kitchenpos.menus.tobe.domain.entity.Menu;
 import kitchenpos.menus.tobe.dto.MenuChangePriceDto;
 import kitchenpos.menus.tobe.dto.MenuCreateDto;
