+++ src/main/java/kitchenpos/menus/tobe/ui/view/MenuViewModel.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.tobe.ui.view;
 
 import java.util.ArrayList;
 import java.util.List;
-import kitchenpos.menus.tobe.application.query.result.MenuQueryResult;
+import kitchenpos.menus.tobe.query.result.MenuQueryResult;
 import kitchenpos.menus.tobe.domain.entity.Menu;
 
 public class MenuViewModel {
